Use size_t for quadratic root loop index
Agent-Logs-Url: https://github.com/rchristie95/MinkowskiKart/sessions/fe8ab386-d52a-43e8-8ebd-ee96b37acd68

+++ src/relativity/relativity_math.cpp
@@ -160,7 +160,7 @@ btVector3 getRelativisticEmissionRelativePosition(
     };
     btScalar emission_dt = btScalar(0.0f);
     bool found = false;
-    for (int i = 0; i < 2; i++)
+    for (size_t i = 0; i < 2; i++)
     {
         if (roots[i] <= btScalar(0.0f) && roots[i] >= btScalar(-1000.0f))
         {
